ajuste datos.
