+++ resteasy-min/README.md
@@ -1,0 +1,8 @@
+# jee8-jax-rs
+Pruebas con jax-rs
+
+http://localhost:8080/resteasy-min/root-path/helloWorld
+http://localhost:8080/resteasy-min/root-path/library/books
+http://localhost:8080/resteasy-min/root-path/library/book/123?name=test
+
+
