--- resteasy-min/src/main/java/es/fjqueal/HelloWorld.java
@@ -1,15 +1,0 @@
-package es.fjqueal;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
- 
-@Path("helloWorld")
-public class HelloWorld {    
-    @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    public String hello() {
-        return "Hello world!";
-    }
-}
